+++ tests/types.py
@@ -12,8 +12,13 @@ class Identifier(BaseModel):
     value: str
 
 
+class Reference(BaseModel):
+    reference: str
+
+
 class Patient(BaseModel):
     resourceType: Literal["Patient"] = "Patient"  # noqa: N815
     id: Union[str, None] = None
     name: list[HumanName]
     identifier: list[Identifier]
+    managingOrganization: Union[Reference, None] = None  # noqa: N815